Update README.md
This repository contains the completed assignment for the Business Analyst position, focusing on the design of an commerce portal for sports equipment. The goal of the project is to outline the major features of the platform for both end customers and vendors, as well as define the user journeys for both customer and vendor onboarding.